pyproject.toml
[project]
name = "ai-courses"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.0.2",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "manim>=0.20.1",
    "moviepy>=2.2.1",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "pillow>=11.0.0",
    "pre-commit>=4.5.1",
    "piper-tts>=1.4.1",
    "qwen-tts>=0.1.0",
    "torch>=2.3.0",
    "pypdfium2>=4.30.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.15.4",
]

[tool.pytest.ini_options]
markers = [
    "integration: end-to-end tests that download and run the TTS model locally",
]

[tool.ruff.lint.per-file-ignores]
# Manim scenes use `from manim import *` by convention
".cache/scripts/lesson.py" = ["F403", "F405"]
"output/*/lesson.py" = ["F403", "F405"]
